# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sanctum API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/meetings": {
      "get": {
        "tags": [
          "meetings"
        ],
        "operationId": "meetings.listMeetings",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "provisional",
                "active",
                "closing",
                "closed",
                "interrupted"
              ]
            },
            "required": false
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "a string at most 512 character(s) long",
              "title": "maxLength(512)",
              "minLength": 1,
              "maxLength": 512
            },
            "required": false,
            "description": "a string at most 512 character(s) long"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NumberFromString"
            },
            "required": false
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "next_cursor"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "workspace_id",
                          "state",
                          "title",
                          "started_at",
                          "ended_at",
                          "timezone",
                          "boundary_revision",
                          "visibility",
                          "processing"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "a Universally Unique Identifier",
                            "format": "uuid",
                            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                          },
                          "workspace_id": {
                            "type": "string",
                            "description": "a Universally Unique Identifier",
                            "format": "uuid",
                            "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "provisional",
                              "active",
                              "closing",
                              "closed",
                              "interrupted"
                            ]
                          },
                          "title": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "started_at": {
                            "type": "string",
                            "description": "a string matching the pattern ^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$",
                            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$"
                          },
                          "ended_at": {
                            "anyOf": [
                              {
                                "type": "string",
                                "description": "a string matching the pattern ^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$",
                                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(\\.\\d{1,6})?Z$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "timezone": {
                            "type": "string"
                          },
                          "boundary_revision": {
                            "type": "integer",
                            "description": "a positive number",
                            "title": "positive",
                            "exclusiveMinimum": 0
                          },
                          "visibility": {
                            "type": "string",
                            "enum": [
                              "restricted",
                              "workspace"
                            ]
                          },
                          "processing": {
                            "type": "object",
                            "required": [
                              "transcript",
                              "notes",
                              "memory",
                              "recording"
                            ],
                            "properties": {
                              "transcript": {
                                "type": "string",
                                "enum": [
                                  "pending",
                                  "partial",
                                  "complete",
                                  "failed"
                                ]
                              },
                              "notes": {
                                "type": "string",
                                "enum": [
                                  "pending",
                                  "partial",
                                  "complete",
                                  "failed"
                                ]
                              },
                              "memory": {
                                "type": "string",
                                "enum": [
                                  "pending",
                                  "partial",
                                  "complete",
                                  "failed"
                                ]
                              },
                              "recording": {
                                "type": "string",
                                "enum": [
                                  "pending",
                                  "partial",
                                  "complete",
                                  "failed"
                                ]
                              }
                            },
                            "additionalProperties": false
                          }
                        },
                        "additionalProperties": false
                      }
                    },
                    "next_cursor": {
                      "anyOf": [
                        {
                          "type": "string",
                          "description": "a string at most 512 character(s) long",
                          "title": "maxLength(512)",
                          "minLength": 1,
                          "maxLength": 512
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The request did not match the expected schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpApiDecodeError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthenticated"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "404": {
            "description": "NotFound",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            }
          },
          "409": {
            "description": "RevisionConflict",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RevisionConflict"
                    },
                    {
                      "$ref": "#/components/schemas/HashConflict"
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unavailable"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HttpApiDecodeError": {
        "type": "object",
        "required": [
          "issues",
          "message",
          "_tag"
        ],
        "properties": {
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Issue"
            }
          },
          "message": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": [
              "HttpApiDecodeError"
            ]
          }
        },
        "additionalProperties": false,
        "description": "The request did not match the expected schema"
      },
      "Issue": {
        "type": "object",
        "required": [
          "_tag",
          "path",
          "message"
        ],
        "properties": {
          "_tag": {
            "type": "string",
            "enum": [
              "Pointer",
              "Unexpected",
              "Missing",
              "Composite",
              "Refinement",
              "Transformation",
              "Type",
              "Forbidden"
            ],
            "description": "The tag identifying the type of parse issue"
          },
          "path": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyKey"
            },
            "description": "The path to the property where the issue occurred"
          },
          "message": {
            "type": "string",
            "description": "A descriptive message explaining the issue"
          }
        },
        "additionalProperties": false,
        "description": "Represents an error encountered while parsing a value to match the schema"
      },
      "PropertyKey": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "object",
            "required": [
              "_tag",
              "key"
            ],
            "properties": {
              "_tag": {
                "type": "string",
                "enum": [
                  "symbol"
                ]
              },
              "key": {
                "type": "string"
              }
            },
            "additionalProperties": false,
            "description": "an object to be decoded into a globally shared symbol"
          }
        ]
      },
      "Unauthenticated": {
        "type": "object",
        "required": [
          "code",
          "retryable",
          "message",
          "_tag"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "unauthenticated"
            ]
          },
          "retryable": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": [
              "Unauthenticated"
            ]
          }
        },
        "additionalProperties": false
      },
      "Forbidden": {
        "type": "object",
        "required": [
          "code",
          "retryable",
          "message",
          "_tag"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "forbidden"
            ]
          },
          "retryable": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "required_scope": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": [
              "Forbidden"
            ]
          }
        },
        "additionalProperties": false
      },
      "NotFound": {
        "type": "object",
        "required": [
          "code",
          "retryable",
          "message",
          "_tag"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "not_found"
            ]
          },
          "retryable": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": [
              "NotFound"
            ]
          }
        },
        "additionalProperties": false
      },
      "RevisionConflict": {
        "type": "object",
        "required": [
          "code",
          "retryable",
          "message",
          "current_revision",
          "_tag"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "revision_conflict"
            ]
          },
          "retryable": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "current_revision": {
            "type": "integer",
            "description": "a non-negative number",
            "title": "nonNegative",
            "minimum": 0
          },
          "_tag": {
            "type": "string",
            "enum": [
              "RevisionConflict"
            ]
          }
        },
        "additionalProperties": false
      },
      "HashConflict": {
        "type": "object",
        "required": [
          "code",
          "retryable",
          "message",
          "existing_sha256",
          "_tag"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "hash_conflict"
            ]
          },
          "retryable": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "existing_sha256": {
            "type": "string"
          },
          "_tag": {
            "type": "string",
            "enum": [
              "HashConflict"
            ]
          }
        },
        "additionalProperties": false
      },
      "Unavailable": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "_tag"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "unavailable"
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "retry_after_ms": {
            "type": "integer",
            "description": "a non-negative number",
            "title": "nonNegative",
            "minimum": 0
          },
          "_tag": {
            "type": "string",
            "enum": [
              "Unavailable"
            ]
          }
        },
        "additionalProperties": false
      },
      "NumberFromString": {
        "type": "string",
        "description": "a string to be decoded into a number"
      },
...
    },
    "securitySchemes": {}
  },
  "security": [],
...
}